requirements.txt
apache-airflow-providers-http
apache-airflow-providers-postgres
requests
psycopg2-binary